Start dividing sync chain : first supplier method
